requirements.txt
streamlit>=1.39,<2.0
openai>=1.30.0,<2.0
python-dotenv>=1.0.1,<2.0
requests>=2.31.0,<3.0
beautifulsoup4>=4.12.0,<5.0
